requirements.txt
kavenegar
django-cors-headers
django
djangorestframework
markdown
django-filter